requirements.txt
pandas
requests
mysql-connector-python
beautifulsoup4
python-dotenv
tmdbsimple
schedule